Profile page changes

+++ Models/PersonModel.cs
@@ -5,7 +5,7 @@
         public int UserId { get; set; }
         public string UserName { get; set; }
         public string? ImageUrl { get; set; }
-        public double PokerBalance { get; set; }
+        public decimal PokerBalance { get; set; }
         // 0 is Pi Lamb, 1 is Hometown, 2 is Both
         public int? FriendGroup { get; set; }
